Fix compile errors and warning messages produces by gcc 4.3.1. (mainly #include's, and some suggestions about using brackets in bool operations)

+++ AMEGIC++/Amplitude/Super_Amplitude.C
@@ -15,7 +15,7 @@ Super_Amplitude::Super_Amplitude(int* _b,int _n,Basic_Sfuncs* _BS,
 
 Super_Amplitude::~Super_Amplitude()                   
 {
-  for (int i=0;i<graphs.size();i++) if (graphs[i]->IsGroup()) delete graphs[i];
+  for (size_t i=0;i<graphs.size();i++) if (graphs[i]->IsGroup()) delete graphs[i];
   graphs.clear();  
 }
